chore: move AddressesData
