+++ bc_obps/registration/api/utils/__init__.py
@@ -1,2 +1,1 @@
-from .operator_utils import *
 from .parent_operator_utils import *
