--- bc_obps/registration/api/utils/operator_utils.py
@@ -1,14 +1,0 @@
-from typing import Optional
-
-
-# ADDRESS HELPERS
-class AddressesData:
-    physical_street_address: str
-    physical_municipality: str
-    physical_province: str
-    physical_postal_code: str
-    mailing_address_same_as_physical: bool
-    mailing_street_address: Optional[str]
-    mailing_municipality: Optional[str]
-    mailing_province: Optional[str]
-    mailing_postal_code: Optional[str]
